Add the share directory to fix sanitized builds
Referencing this issue on bazelbuild/bazel:

https://github.com/bazelbuild/bazel/issues/10510

The output of `clang --print-resource-dir` needs to be used to determine
the share directory to be added to the include path. Since we can
predict the path, we don't need to run clang to figure it out, but we do
need to add it to cxx_builtin_include_directories.