Add more LLVM versions, including some custom-built 15.0.7 versions (#8)
* Add a bunch of LLVM distributions

* The rest of the changes required to get it to find the right version

* Forcibly select x86_64 for the time being

Sorbet uses the x86_64 version of Bazel under Rosetta when running on
Apple Silicon. This has a weird side effect of tricking Bazel into
thinking that the target architecture is x86_64 when it does things like
invoke the C++ compiler via cc_binary and friends.

But this `llvm_release_name.py` tool just runs with Python on the user's
system. So when we ask it for `platform.machine()` it will truthfully
say `arm64`, which will then attempt to download the arm64 version of
llvm+clang, and that will interfere with our clever trick of forcing
Bazel to always build an x86_64 binary even on Apple Silicon.

Eventually, we will want to do the work to support building on Apple
Silicon, but that's a problem for another day.

* Need another cxx_builtin_include_directories

This is already present upstream, though based on the git history I
could only find that the change was made after the move to the new
toolchain format:

    https://github.com/bazel-contrib/toolchains_llvm/commit/4e6d9960a24560765dc6557861ee34b74e60db92

4e6d9960a24560765dc6557861ee34b74e60db92

* Add another URL

* Add another URL