Use LLD linker for Darwin
Originally this commit:

<https://github.com/bazel-contrib/toolchains_llvm/pull/286/commits/8179a799053e56f312afda027afe6ac59ad2ab31>

The PR that this change landed in ended up making a followup commit to
make the change optional, so that people could opt into it by passing
`--linkopt` and/or `--host_linkopt` to set something like

    --linkopt=-fuse-ld=ld64.lld

We're seeing behavior where if we **ever** use the system linker, it
fails to link (because it attempts to link libunwind at the wrong
location), so we need to make this linker be used all the time. There is
no way to opt out of using `ld64.lld` with this commit.